+++ Emaildataplane/EmailDPClient.cs
@@ -103,7 +103,7 @@ namespace Oci.EmaildataplaneService
                     ServiceName = "EmailDP",
                     OperationName = "SubmitEmail",
                     RequestEndpoint = $"{method.Method} {requestMessage.RequestUri}",
-                    ApiReferenceLink = "",
+                    ApiReferenceLink = "https://docs.oracle.com/iaas/api/#/en/emaildeliverysubmission/20220926/EmailSubmittedResponse/SubmitEmail",
                     UserAgent = this.GetUserAgent()
                 };
                 this.restClient.CheckHttpResponseMessage(requestMessage, responseMessage, apiDetails);