+++ Stackmonitoring/models/MonitoredResource.cs
@@ -143,6 +143,22 @@ namespace Oci.StackmonitoringService.Models
         public System.Nullable<LicenseType> License { get; set; }
         
         /// <value>
+        /// Source type to indicate if the resource is stack monitoring discovered, OCI native resource, etc.
+        /// 
+        /// </value>
+        [JsonProperty(PropertyName = "sourceType")]
+        [JsonConverter(typeof(Oci.Common.Utils.ResponseEnumConverter))]
+        public System.Nullable<SourceType> SourceType { get; set; }
+        
+        /// <value>
+        /// Resource Category to indicate the kind of resource type.
+        /// 
+        /// </value>
+        [JsonProperty(PropertyName = "resourceCategory")]
+        [JsonConverter(typeof(Oci.Common.Utils.ResponseEnumConverter))]
+        public System.Nullable<ResourceCategory> ResourceCategory { get; set; }
+        
+        /// <value>
         /// List of monitored resource properties.
         /// 
         /// </value>