+++ CHANGELOG.md
@@ -4,6 +4,24 @@ All notable changes to this project will be documented in this file.
 
 The format is based on Keep a [Changelog](http://keepachangelog.com/).
 
+## 85.0.0 - 2024-04-09
+### Added
+- Support for calling Oracle Cloud Infrastructure services in the ap-dcc-gazipur-1 region
+- Support for the DNS-based backend health check in the Network Load Balancer service
+- Support for Fail Open in the Network Load Balancer service
+- Support for adding and updating Instant failover in the Network Load Balancer service
+- Support for adding and updating source type and resource category for resource types in the Stack Monitoring service
+- Support for searching resources based on resource category,  source type, multiple compartments, multiple lifecycle states in the Stack Monitoring service
+- Support for filtering listed resources based on lifecycle status in the Stack Monitoring service
+- Support for creating tasks with new config parameters in the Stack Monitoring service
+- Support for Composite Resource Principal integration in the Data Flow service
+
+### Changed
+- The `TargetFramework` of projects `Objectstoragetests`, `Commontests` and `Examples` was changed from `netcoreapp3.1` to `net6.0`
+
+### Breaking Changes
+- The Application Migration service was decommissioned and removed
+
 ## 84.0.0 - 2024-04-02
 ### Added
 - Support for assigned private IP and single stack IPV6 feature for Network Load Balancer Service